auth: Move BasicAuthConfig to its own file
